Refactor: Split crash logic out of Vehicle.cpp
Part of #25951

This moves all crash and collision functions from Vehicle.cpp into a dedicated Vehicle.Crash.cpp file. Functions are moved intact with
no logic changes. Also, ride_get_train_index_from_vehicle is made non static and it's now declared in Vehicle.h as it's now needed across translation units
