+++ src/openrct2/libopenrct2.vcxproj
@@ -1122,6 +1122,7 @@
     <ClCompile Include="ride\TrackStyle.cpp" />
     <ClCompile Include="ride\TrainManager.cpp" />
     <ClCompile Include="ride\Vehicle.BoatHire.cpp" />
+    <ClCompile Include="ride\Vehicle.Crash.cpp" />
     <ClCompile Include="ride\Vehicle.cpp" />
     <ClCompile Include="ride\VehicleData.cpp" />
     <ClCompile Include="ride\VehicleGeometry.cpp" />
